Move AppContext to middleware
