+++ src/FunderMaps.AspNetCore/Middleware/AspAppContextMiddleware.cs
@@ -4,7 +4,7 @@ using FunderMaps.Core.Authentication;
 using FunderMaps.Core.Entities;
 using Microsoft.AspNetCore.Http;
 
-namespace FunderMaps.AspNetCore.Components
+namespace FunderMaps.AspNetCore.Middleware
 {
     /// <summary>
     ///     Create the <see cref="Core.AppContext"/> from the <see cref="HttpContext"/>.
